added mappers to ts http lib
