+++ lib/typescript/httpclient/endpoints/connectChannel.ts
@@ -1,27 +1,8 @@
 import {doFetchFromBackend} from '../api';
-import {ConnectChannelRequestPayload, ConnectChannelRequestApiPayload} from '../payload';
-import {Channel} from '../model';
+import {ConnectChannelRequestPayload} from '../payload';
 import {ChannelApiPayload} from '../payload/ChannelApiPayload';
-
-const connectChannelApiMapper = (payload: ConnectChannelRequestPayload): ConnectChannelRequestApiPayload => {
-  return {
-    source: payload.source,
-    source_channel_id: payload.sourceChannelId,
-    token: payload.token,
-    name: payload.name,
-    image_url: payload.imageUrl,
-  };
-};
-
-const channelMapper = (payload: ChannelApiPayload): Channel => {
-  return {
-    name: payload.name,
-    source: payload.source,
-    sourceChannelId: payload.source_channel_id,
-    imageUrl: payload.image_url,
-    connected: true,
-  };
-};
+import {connectChannelApiMapper} from '../mappers/connectChannelApiMapper';
+import {channelMapper} from '../mappers/channelMapper';
 
 export function connectChannel(requestPayload: ConnectChannelRequestPayload) {
   return doFetchFromBackend('channels.connect', connectChannelApiMapper(requestPayload))
