+++ lib/typescript/httpclient/endpoints/exploreChannels.ts
@@ -1,18 +1,7 @@
 import {doFetchFromBackend} from '../api';
 import {ExploreChannelRequestPayload} from '../payload';
-import {Channel} from '../model';
 import {ChannelsPayload} from '../payload/ChannelsPayload';
-
-const channelsMapper = (payload: ChannelsPayload, source?: string): Channel[] => {
-  return payload.data.map(
-    (entry: Channel): Channel => {
-      return {
-        source,
-        ...entry,
-      };
-    }
-  );
-};
+import {channelsMapper} from '../mappers/channelsMapper';
 
 export function exploreChannels(requestPayload: ExploreChannelRequestPayload) {
   return doFetchFromBackend('channels.explore', requestPayload)
