+++ lib/typescript/httpclient/endpoints/listChannels.ts
@@ -1,17 +1,6 @@
 import {doFetchFromBackend} from '../api';
-import {Channel} from '../model';
 import {ChannelsPayload} from '../payload/ChannelsPayload';
-
-const channelsMapper = (payload: ChannelsPayload, source?: string): Channel[] => {
-  return payload.data.map(
-    (entry: Channel): Channel => {
-      return {
-        source,
-        ...entry,
-      };
-    }
-  );
-};
+import {channelsMapper} from '../mappers/channelsMapper';
 
 export function listChannels() {
   return doFetchFromBackend('channels.list')
