+++ lib/typescript/httpclient/endpoints/listTags.ts
@@ -1,17 +1,6 @@
 import {doFetchFromBackend} from '../api';
-import {Tag} from '../model';
 import {ListTagsResponsePayload} from '../payload';
-
-const tagMapper = {
-  BLUE: 'tag-blue',
-  RED: 'tag-red',
-  GREEN: 'tag-green',
-  PURPLE: 'tag-purple',
-};
-
-const tagsMapper = (serverTags: Tag[]): Tag[] => {
-  return serverTags.map(t => ({id: t.id, name: t.name, color: tagMapper[t.color] || 'tag-blue'}));
-};
+import {tagsMapper} from '../mappers/tagsMapper';
 
 export function listTags() {
   return doFetchFromBackend('tags.list')
