+++ lib/typescript/httpclient/mappers/channelsMapper.ts
@@ -1,0 +1,13 @@
+import {ChannelsPayload} from '../payload/ChannelsPayload';
+import {Channel} from '../model';
+
+export const channelsMapper = (payload: ChannelsPayload, source?: string): Channel[] => {
+  return payload.data.map(
+    (entry: Channel): Channel => {
+      return {
+        source,
+        ...entry,
+      };
+    }
+  );
+};
