+++ lib/typescript/httpclient/mappers/connectChannelApiMapper.ts
@@ -1,0 +1,11 @@
+import {ConnectChannelRequestPayload, ConnectChannelRequestApiPayload} from '../payload';
+
+export const connectChannelApiMapper = (payload: ConnectChannelRequestPayload): ConnectChannelRequestApiPayload => {
+  return {
+    source: payload.source,
+    source_channel_id: payload.sourceChannelId,
+    token: payload.token,
+    name: payload.name,
+    image_url: payload.imageUrl,
+  };
+};
