+++ lib/typescript/httpclient/mappers/disconnectChannelApiMapper.ts
@@ -1,0 +1,9 @@
+import {DisconnectChannelRequestPayload, DisconnectChannelRequestApiPayload} from '../payload';
+
+export const disconnectChannelApiMapper = (
+  payload: DisconnectChannelRequestPayload
+): DisconnectChannelRequestApiPayload => {
+  return {
+    channel_id: payload.channelId,
+  };
+};
