+++ lib/typescript/httpclient/mappers/userMapper.ts
@@ -1,0 +1,12 @@
+import {UserPayload} from '../payload/UserPayload';
+import {User} from '../model';
+
+export const userMapper = (payload: UserPayload): User => {
+  return {
+    id: payload.id,
+    firstName: payload.first_name,
+    lastName: payload.last_name,
+    displayName: payload.first_name + ' ' + payload.last_name,
+    token: payload.token,
+  };
+};
